// domain/osekai/enums/OsekaiRanking.enum.ts
export enum EOsekaiRanking {
    MedalCount = "MedalCount",
    MedalRarity = "MedalRarity",

    PPStandardDeviation = "PPStandardDeviation",
    PPTotal = "PPTotal",

    LevelStandardDeviation = "LevelStandardDeviation",
    LevelTotal = "LevelTotal",

    AccuracyStandardDeviation = "AccuracyStandardDeviation",
    AccuracyTotal = "AccuracyTotal",

    Replays = "Replays",

    RankedMapsets = "RankedMapsets",
    LovedMapsets = "LovedMapsets",

    Subscribers = "Subscribers",
    Badges = "Badges",

    PlaytimeStandardDeviation = "PlaytimeStandardDeviation",
    PlaytimeTotal = "PlaytimeTotal",
}

export enum EOsekaiAllModeRankingType {
    StandardDeviation = "StandardDeviation",
    Total = "Total",
}

export enum EOsekaiMapsetRankingType {
    Ranked = "Ranked",
    Loved = "Loved",
}

export enum EOsekaiRankingEntryType {
    User = "User",
    Medal = "Medal",
}

export enum EOsekaiRankingValueFormat {
    Integer = "Integer",
    Decimal = "Decimal",
    PP = "PP",
    MedalRarity = "MedalRarity",
    PlaytimeTotal = "PlaytimeTotal",
    PlaytimeStandardDeviation = "PlaytimeStandardDeviation",
}